+++ src/lib.rs
@@ -1,0 +1,2 @@
+/// Export the cli module
+pub mod cli;
