+++ tests/blocks.rs
@@ -1,0 +1,28 @@
+// use std::{str::FromStr, path::PathBuf};
+
+// use reth_primitives::{rpc::H256, U256};
+
+// use op_reth::cli::db;
+// use op_reth::cli::blocks;
+
+// const BLOCKS_PATH: &str = "data/export_0_4061223";
+
+#[test]
+fn test_blocks_from_file() {
+    // let blocks = blocks::ErigonBlock::from_file(BLOCKS_PATH).unwrap();
+    // assert_eq!(0, receipts[0].ty);
+    // assert_eq!(1, receipts[0].status);
+    // assert_eq!(151191, receipts[0].cumulative_gas_used);
+    // assert_eq!(H256::from_str("
+    // 0x7334ddc1f6beaf66892c25cffdecec275cdfabaf4def047f0c3ce20e6f6483e8").unwrap(),
+    // receipts[0].tx_hash); assert_eq!(151191, receipts[0].gas_used);
+    // assert_eq!(H256::from_str("
+    // 0x15d55041e8f7b0d1f303b6d4cefe2d2efc257d67acd9f17307261a8f7d786e0e").unwrap(),
+    // receipts[0].block_hash); assert_eq!(U256::from(1), receipts[0].block_number);
+    // assert_eq!(0, receipts[0].transaction_index);
+    // assert_eq!(U256::from(1), receipts[0].l1_gas_price);
+    // assert_eq!(U256::from_str("0x1b62").unwrap(), receipts[0].l1_gas_used);
+    // assert_eq!(U256::from_str("0x2913").unwrap(), receipts[0].l1_fee);
+    // assert_eq!("1.5", receipts[0].l1_fee_scalar);
+    // assert_eq!(4029549, receipts.len());
+}
